Correcciones mayor contable, quitar restriccion modulo proveedor

+++ src/main/com/encens/khipus/service/accouting/VoucherAccoutingServiceBean.java
@@ -473,25 +473,25 @@ public class VoucherAccoutingServiceBean extends GenericServiceBean implements V
     public Double getBalance(Date startDate, String cashAccountCode){
 
         try {
-            Double balanceD = ((BigDecimal) em.createQuery("select sum(voucherDetail.debit) from VoucherDetail voucherDetail " +
+            BigDecimal balanceD = (BigDecimal) em.createQuery("select sum(voucherDetail.debit) from VoucherDetail voucherDetail " +
                             " join voucherDetail.voucher voucher " +
                             " where voucher.date < :startdate " +
                             " and voucherDetail.account = :cashAccountCode " +
                             " and voucher.state <> 'ANL' ")
                     .setParameter("startdate", startDate)
                     .setParameter("cashAccountCode", cashAccountCode)
-                    .getSingleResult()).doubleValue();
+                    .getSingleResult();
 
-            Double balanceC = ((BigDecimal) em.createQuery("select sum(voucherDetail.credit) from VoucherDetail voucherDetail " +
+            BigDecimal balanceC = (BigDecimal) em.createQuery("select sum(voucherDetail.credit) from VoucherDetail voucherDetail " +
                             " join voucherDetail.voucher voucher " +
                             " where voucher.date < :startdate " +
                             " and voucherDetail.account = :cashAccountCode " +
                             " and voucher.state <> 'ANL' ")
                     .setParameter("startdate", startDate)
                     .setParameter("cashAccountCode", cashAccountCode)
-                    .getSingleResult()).doubleValue();
+                    .getSingleResult();
 
-            return (balanceD != null ? balanceD : 0.0) - (balanceC != null ? balanceC : 0.0);
+            return (balanceD != null ? balanceD.doubleValue() : 0.0) - (balanceC != null ? balanceC.doubleValue() : 0.0);
         } catch (NoResultException e) {
             return 0.0;
         }
